+++ radiant_mlhub/models.py
@@ -6,12 +6,15 @@ from collections.abc import Sequence
 from copy import deepcopy
 from enum import Enum
 from pathlib import Path
-from typing import Iterator, List, Optional, Union
+from typing import Iterable, Iterator, List, Optional, Union
 
 import pystac
 
 from . import client
 from .exceptions import EntityDoesNotExist
+
+TagOrTagList = Union[str, Iterable[str]]
+TextOrTextList = Union[str, Iterable[str]]
 
 
 class Collection(pystac.Collection):
@@ -389,13 +392,17 @@ class Dataset:
         return self._collections
 
     @classmethod
-    def list(cls, **session_kwargs) -> List['Dataset']:
+    def list(cls, *,  tags: Optional[TagOrTagList] = None, text: Optional[TextOrTextList] = None, **session_kwargs) -> List['Dataset']:
         """Returns a list of :class:`Dataset` instances for each datasets hosted by MLHub.
 
         See the :ref:`Authentication` documentation for details on how authentication is handled for this request.
 
         Parameters
         ----------
+        tags : A list of tags to filter datasets by. If not ``None``, only datasets containing all
+            provided tags will be returned.
+        text : A list of text phrases to filter datasets by. If not ``None``, only datasets
+            containing all phrases will be returned.
         **session_kwargs
             Keyword arguments passed directly to :func:`~radiant_mlhub.session.get_session`
 
@@ -405,7 +412,7 @@ class Dataset:
         """
         return [
             cls(**d, **session_kwargs)
-            for d in client.list_datasets(**session_kwargs)
+            for d in client.list_datasets(tags=tags, text=text, **session_kwargs)
         ]
 
     @classmethod